+++ .rubocop.yml
@@ -1,5 +1,5 @@
 AllCops:
-  TargetRubyVersion: 3.0
+  TargetRubyVersion: 3.0.1
   Exclude:
     - 'tmp/**/*'
     - "spec/**/*"
