+++ lesson8_vera/1_keywords.rb
@@ -1,20 +1,6 @@
 # frozen_string_literal: true
+require_relative 'lib/keywords'
 
-require 'set'
-
-class Keywords
-  @@keywords = Set.new
-
-  def initialize(keyword)
-    keyword.each do |w|
-      @@keywords.add w.downcase
-    end
-  end
-
-  def self.words
-    @@keywords
-  end
-end
 [
   ['праздник', 'парад', 'День независимости'],
   %w[феерверки праздник концерт],
