+++ lesson8_vera/2_ships.rb
@@ -1,0 +1,6 @@
+require 'ostruct'
+ns1 = OpenStruct.new(torpedoes: 3, rocket: 9)
+ns2 = OpenStruct.new(torpedoes: 11, rocket: 2)
+ns3 = OpenStruct.new(torpedoes: 7, rocket: 12)
+
+ns1.torpedoes
